# src/views/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MediaFlow CRM API",
    "description": "Complete API documentation for MediaFlow CRM - A comprehensive production management system for media companies. Manage contacts, projects, tasks, team members, and file uploads with a modern RESTful API.",
    "version": "1.0.0",
    "contact": {
      "name": "MediaFlow Support",
      "email": "support@mediaflow.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.mediaflow.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/team/members": {
      "get": {
        "tags": ["Team"],
        "summary": "List team members",
        "description": "Get paginated list of team members",
        "operationId": "listTeamMembers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          },
          {
            "name": "role",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["admin", "manager", "member", "viewer"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "inactive", "on_leave"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Team members retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Team"],
        "summary": "Create team member",
        "description": "Create a new team member (auto-generates password)",
        "operationId": "createTeamMember",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMemberInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team member created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberResponse"
                }
              }
            }
          }
        }
      }
    },
    "/team/members/{id}": {
      "get": {
        "tags": ["Team"],
        "summary": "Get team member",
        "description": "Get team member details with assignments",
        "operationId": "getTeamMember",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Team member retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Team"],
        "summary": "Update team member",
        "description": "Update team member information",
        "operationId": "updateTeamMember",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamMemberInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team member updated successfully"
          }
        }
      },
      "delete": {
        "tags": ["Team"],
        "summary": "Deactivate team member",
        "description": "Deactivate a team member (soft delete)",
        "operationId": "deactivateTeamMember",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Team member deactivated successfully"
          }
        }
      }
    },
...
    "/team/assignments/{id}": {
      "delete": {
        "tags": ["Team"],
        "summary": "Remove assignment",
        "description": "Remove project assignment",
        "operationId": "removeAssignment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Assignment removed successfully"
          }
        }
      }
    },
    "/team": {
      "get": {
        "tags": ["Team"],
        "summary": "List teams",
        "description": "Get list of all teams",
        "operationId": "listTeams",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Teams retrieved successfully"
          }
        }
      },
      "post": {
        "tags": ["Team"],
        "summary": "Create team",
        "description": "Create a new team",
        "operationId": "createTeam",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team created successfully"
          }
        }
      }
    },
    "/team/{id}": {
      "get": {
        "tags": ["Team"],
        "summary": "Get team",
        "description": "Get team details with members",
        "operationId": "getTeam",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Team retrieved successfully"
          }
        }
      },
      "put": {
        "tags": ["Team"],
        "summary": "Update team",
        "description": "Update team information",
        "operationId": "updateTeam",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Team updated successfully"
          }
        }
      },
      "delete": {
        "tags": ["Team"],
        "summary": "Delete team",
        "description": "Delete a team",
        "operationId": "deleteTeam",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Team deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer {token}"
      }
    },
    "parameters": {
      "IdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Resource ID (UUID)",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "Page number",
        "schema": {
          "type": "integer",
          "default": 1,
          "minimum": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "Items per page",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        }
      },
...
    },
    "schemas": {
...
      "TeamMemberInput": {
        "type": "object",
        "required": ["first_name", "last_name", "email"],
        "properties": {
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "manager", "member", "viewer"]
          },
          "position": {
            "type": "string"
          },
          "department": {
            "type": "string"
          }
        }
      },
      "TeamMemberResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        }
      },
      "TeamMemberListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPrevPage": {
            "type": "boolean"
          }
        }
      }
    },
...
  }
}